+++ tomcat/src/main/java/jakarta/http/ContentType.java
@@ -1,4 +1,4 @@
-package nextstep.jwp.http;
+package jakarta.http;
 
 import java.util.Arrays;
