+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -5,8 +5,8 @@ import java.io.IOException;
 import java.io.InputStreamReader;
 import java.net.Socket;
 import nextstep.jwp.exception.UncheckedServletException;
-import nextstep.jwp.http.reqeust.HttpRequest;
-import nextstep.jwp.http.response.HttpResponse;
+import jakarta.http.reqeust.HttpRequest;
+import jakarta.http.response.HttpResponse;
 import org.apache.catalina.core.controller.ControllerContainer;
 import org.apache.coyote.Processor;
 import org.slf4j.Logger;
